Trying to save

+++ src/Bld.LibcameraNet/MappedPlane.cs
@@ -4,5 +4,5 @@ public class MappedPlane
 {
     public int fd { get; set; }
     public nuint offset { get; set; }
-    public nuint len { get; set; }
+    public UInt64 len { get; set; }
 }
